Feat: jar

+++ src/main/java/com/example/security/repository/WordRepository.java
@@ -1,20 +1,19 @@
 package com.example.security.repository;
 
-import com.example.security.model.Note;
 import com.example.security.model.Word;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface WordRepository extends JpaRepository<Word, Integer> {
+public interface WordRepository extends JpaRepository<Word, Long> {
 
     @Query(value =
         "SELECT * " +
             "FROM word " +
             "WHERE note_id = :noteId",
         nativeQuery = true)
-    List<Word> findByNoteId(@Param("noteId") Integer noteId);
+    List<Word> findByNoteId(@Param("noteId") Long noteId);
 
     @Query(value =
         "SELECT * " +
